+++ code/crates/core-consensus/src/params.rs
@@ -11,12 +11,6 @@ pub use malachitebft_core_driver::ThresholdParams;
 /// Consensus parameters.
 #[derive_where(Clone, Debug)]
 pub struct Params<Ctx: Context> {
-    /// The initial height
-    pub initial_height: Ctx::Height,
-
-    /// The initial validator set
-    pub initial_validator_set: Ctx::ValidatorSet,
-
     /// The address of this validator
     pub address: Ctx::Address,